Main folder cleanup and Image update
Removd some classes I added to help make the end screen. Readded the images for the home and splash screen, as I had set them to null in my build (as assets was not yet added). Added support for a background image in the game screen and the end screen (null right now)